requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.4
python-dateutil
reportlab==4.4.6
